Move dialogs from layout to index page - Add stub settings dialog
